dashboard/requirements.txt
confluent-kafka>=2.6.1
flask>=3.1.0
kafka-python>=2.0.2
omegaconf>=2.3.0
hydra-core>=1.3.2
requests>=2.32.3
PyYAML>=6.0
docker>=6.0.0
psutil>=7.0.0